Fix linting errors and move the reference to not be linted
