+++ src/components/landing/landing.js
@@ -1,5 +1,7 @@
 import React from "react";
 
 const Landing = props => {
-	return <p>Hello, World!</p>
+	return <p>Hello, World!</p>;
 };
+
+export default Landing;
